requirements.txt
python-telegram-bot
pandas
gspread
oauth2client
python-dotenv
